backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
yfinance
requests
